- Jorrit optimized the kdtree a bit so that it now does 20 attempts 	  instead of 10 to find the best split location. Also the cost is 	  now calculated based on a measure that looks more at number of 	  cut objects as opposed to balanced tree.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@15903 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csgeom/kdtree.h
@@ -183,7 +183,7 @@ private:
    * If this function returns a negative quality this means the
    * split should not be performed at all.
    */
-  int FindBestSplitLocation (int axis, float& split_loc);
+  float FindBestSplitLocation (int axis, float& split_loc);
 
   /**
    * If this node is a leaf then we will split the objects currently
